feat: add loan and deposit stats dto files #51

+++ server/src/site/doget/dto/raw/DepositInfoRawDto.java
@@ -10,8 +10,8 @@ public class DepositInfoRawDto {
     private String accSbjcNm;
     private String custDscdNm;
     private Integer cnt;
-    private Integer max;
-    private Integer avg;
-    private Integer mid;
-    private Integer mod;
+    private Long max;
+    private Long avg;
+    private Long mid;
+    private Long mod;
 }
